Workaround inability to download and upload CDN files
