+++ src/platform/logging/types.ts
@@ -45,5 +45,6 @@ export enum TraceOptions {
 }
 
 export enum ConsoleForegroundColors {
-    Green = '\x1b[32m'
+    Green = '\x1b[32m',
+    Red = `\x1b[31m`
 }
